Add old tests to new test_behavior file
